backend/requirements.txt
Django==3.2.3
django_filter==23.2
djangorestframework==3.12.4
django-colorfield==0.10.1
django-cors-headers==3.13.0
djoser==2.1.0
drf-extra-fields==3.4.0
flake8==6.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.3
pillow==9.3.0
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.4.4
